init fix

+++ init.sql
@@ -1,7 +1,7 @@
 create table users (
-    tg_id          integer unique primary key,
+    tg_id          bigint  unique primary key,
     tg_username    text    not null,
     sol_public_key text    not null,
-    tg_chat_id     bigint not null,
+    tg_chat_id     bigint  not null,
     created_at  timestamp without time zone default NOW()
 );
